Move handleTokenClick to chatUtils
